setup update

+++ ingress-nginx/certificate.sh
@@ -9,6 +9,18 @@ if [[ -z $domain ]]; then
 	domain="develop"
 fi
 
+echo "apply minikube ip patch"
+
+IP=$(minikube ip)
+
+if [ -z $IP ]; then
+	echo "cannot get minikube ip"
+	exit 1
+fi
+
+
+sed "s/minikube_ip/$IP/g" $dir/certificate_patch.cnf > $dir/certificate.cnf
+
 echo "create self signed wildcard certificate for $domain"
 
 rm -f wildcard.*
@@ -16,7 +28,7 @@ rm -f wildcard.*
 openssl req -nodes -newkey rsa:2048 -new -sha256 \
   -keyout $dir/wildcard.$domain.key \
   -out $dir/wildcard.$domain.csr \
-  -subj "/O=develop/CN=*.$domain"
+  -config $dir/certificate.cnf  
 
 openssl req -text -in $dir/wildcard.$domain.csr -noout -verify
